+++ src/pages/news/AddNewsPage.tsx
@@ -135,7 +135,7 @@ function AddNewsPage() {
     let fieldapp = false;
     let fieldimg = false;
 
-    if(newsName && newsDescription && newsStatus){
+    if(newsName && (newsDescription != "<p><br></p>") && newsStatus){
       fieldInfo = false
     }
     else{